Ignore C4820, treat GCC/Clang warnings as errors, fix assignment.
The classes TestCase and TestSuite use 10 and 6 bytes, respectively,
of extra space because of padding added by the compiler to ensure the
data in each class is aligned properly, which is negligible given
consideration that even if there were 100 that would add less than 1KB.
It may be a problem if there were 100000 cases which would add ~1MB of
extra space just for proper alignment, which is significant, but if you
have 100k test cases this is probably not the right library you should
be using for testing your code implying very large code size and
significantly-more complexity than this library is intended to test.
The space-time tradeoff in this case is time, and proper natural
alignment is favored for speed instead of using additional CPU cycles
to access class data elements.
I could be misunderstanding alignment in this case, and it could be
something to revisit if it becomes an issue later.

Added -Werror so GCC/Clang warnings are treated as errors; I don't want
to leave warnings hanging around, and not using -Werror allows that.

Fixed a potential issue (thank you MSVC via warning C4365) where a
signed integer was being assigned to an unsigned integer, which could
cause overflow.
This issue is in the test runtime code that records the runtime of each
test case.