Suppress MSVC warning C5039
There's an issue in winbase.h where an API is passing a reference to a
function that could throw an exception to a C API, resulting in
undefined behavior.
TestCPP is purely C++ and is not designed to be used with C, so I think
this is probably a safe warning to suppress.