Fix dtor race condition, add new function to get suite failed case count
Should have realized from the beginning when implementing the stream
capture code that this needed to be per execution not per case.
There is one clog, cout, and cerr, treating them otherwise was an
obvious mistake and was the root cause of the windows crash.
It makes me wonder why on earth it wasn't crashing on nix!
It was just bailing out silently and not finishing the suite runs on
nix but was not crashing the test executable.
Very strange.
So the windows crash is fixed.
Now, it initializes and handles the stream capture with a counting-
semaphore-ish technique where it will support multiple runs and
capturing/uncapturing automatically to ensure control returns to the
suite for the next case.
Dtors, ctors, and assignment operators not reflect this also.
Since they were made static they are referenced through the class
instead of through this.

Added a new API to get the test suite failed test count, which I chose
to use as a return value since 0 failures aligns with the success
return code 0 and anything other than 0 causes the test execution to
register with CTest as a failing test module.

The output capture check function now outputs detail about why a match
failed based on whether the compare option is EXACT or CONTAINS,
taking into account that clog might be captured.