Fix setNotifyPassed test, remove dead code
The debugLog statement I removed will never be called.
Thus removed it.

The setNotifyPassed test must have been succeeding erroneously previous
to fixing the stream capture code.
The test has been reworked to properly check for the output when test
pass notification is enabled and for the absence of it when it is
disabled.