Ensure unsigned values are safely cast to signed values.
Implemented a function based on this StackOverflow page:
https://stackoverflow.com/questions/13150449/efficient-unsigned-to-signed-cast-avoiding-implementation-defined-behavior
This function ensures that when an unsigned value is used as the return
value of a function (in this case the main function) the value is safe
for this use (that it does not overflow the signed value it is
converted to); in this particular case when used as a return value for
the main function.
If the number cannot be safely returned, a runtime_error is thrown.

Applied this to the main functions to fix MSVC warning C4365 which
ensures it's known when unsigned values are implicitly converted to
signed values.