Fix cppcheck args so it properly ignores missing system includes
Currently there are no #include <> statements that point to non-system
includes, except Boost, but I believe it's safe to ignore that as well
in the case of this library.
The Boost include in this case is not necessarily under my control and
contextually is something like a system-include along the lines of the
standard library includes.