Fix static analysis workflow, add cppcheck + info to readme
Also clean up clang-tidy configuration, stop including the deprecated
and removed AnalyzeTemporaryDtors option as the option no longer is
used or even exists in versions of clang-tidy after version 6.

Merge static analysis workflows to use only the debug workflow.
It's not necessary to run the analysis on both release and debug
build configurations, from what I can see.
This is reflected in the README.

Added information to the README on running cppcheck from CodeLite.
Added information to the README on running cppcheck on the project
from the command line/terminal.

Added clang-tidy information to the README.

Added custom build action to the CodeLite project to run cppcheck.