Cleanup, TC member initialization, TC capture bools, fix message bug
Initialize everything in TestCase with sane defaults.
Name the parameter in the logTestFailure function prototype.

Add boolean members that indicate whether for the given TestCase the
output should be captured.
Add implementation details for capture bools.
This ensures all are counted properly.

Rename the testPassedMessage TestSuite member to testSuitePassedMessage.
This properly differentiates it from the similar member of TestCase.

No addTest specializations in header file, only the template.
Moved the specialization from the header file that was there to the
source file.
This fixed an issue when testing where the compiler could not find the
proper specialization.

Flesh out/add/fix docs.

Fixed a huge bug in how the test passage message is enabled/disabled en-
mass through a TestSuite.
Previously the vector was iterated over with the newer style, but this
causes a copy to be made on every iteration of every TestCase, which
is completely not what we want to do here.
Fall back to index-based vector iteration so we're properly modifying
each original TestCase and not making copies thinking we're modifying
the actual TestCases.