Fix multiple frees on streams, make sure test failures aren't silenced
Make sure stream capture does not affect logging of test failures.
-----------------------------------------------------------------------
Created a subroutine for logging test failures where the output stream
is injected instead of just using std::clog.
This allows the TestCase::logTestFailure function to log to the real
clog stream and the captured (if captured) clog stream to ensure any
expected output is present for analysis.
This ensures test failures are logged even if the stream being logged to
is captured for analysis for the given test.
This was an oversight on my part, and there are more than likely other
instances where this will need to be corrected.

Fix multiple frees on streams (stringstreams and streambufs)
-----------------------------------------------------------------------
I've been working on this periodically since last month to figure out
what is happening.
It turns out, std::unique_ptr seems to add an exit handler to the
program (this is a possibility, I haven't looked at the STL
implementation code to confirm) that calls the deleter when it is a
static object.
The default deleter just calls delete on the pointed-to object without
checking that and with no context around the object.
The solution ended up being constructing each std::unique_ptr with a
custom no-op deleter to allow the TestCase destructor to properly
handle desctruction, since proper functionality of the TestCase code
depends on the semaphore behavior defined in the construction and
destruction of TestCases depending on the stream capture arguments,
so it knows when to free the buffers: only when nothing is using them.
The std::unique_ptr template arguments are adjusted accordingly, and
the buffers are now freed at the proper time and place.
The streambufs are also left alone with custom no-op deleters, as we
have no business freeing streambufs that are not ours to delete (the
underlying streams for std::cout, std::clog, and std::cerr).