Clean up how test failures in a suite are recorded; Fixed failing tests
There is now only one place for incrementing, and the conditions are
now logically split and actions clearer.

Tests in TestSuiteTests are no longer failing.
The output buffer was not being cleared and there was lingering buffer
content from previous tests.
The tests were checking that nothing was logged related to test passage,
so the checks for lack of output were failing.

All tests pass! :)