It looks like that bug fix fixed the windows segfault in the TC tests
Removing the debug logs.

Since the std::streambuf pointer returned by cout.rdbuf() result was
already assigned to a std::unique_ptr, it must have been what was
causing the segfault (really an Access Violation, this is Windows not
*nix/BSD).
The weird thing is that the same code was not causing the issue on any
other platform with any other compiler.
The only time the issue popped up was on Windows with the MSVC++ cl
compiler, compiling with debug options, where there were no compile
warnings or errors!
Very strange, but the good news is it pointed out 2 really severe but
subtle bugs of the same kind in 2 different places, where the wrong
std::streambuf was being stored for output capturing, which would have
caused an inability to revert the capture for tests that don't capture
output after the capturing test(s) have finished running.
All because of a copy-paste error 2 times over. Yeesh.