Added information to README about static analysis and planned features
Fixed the order of GH Actions workflows.
Added flawfinder and infer to the list of static analysis tools.
Added that debug and release are also taken into consideration for presets.
Updated the vcpkg and conan sections to more acurately reflect status.
Split the Features section into Current and Planned.
Added a long list of new assertions I plan on adding to the library.