Update CMake with new test suite, fix assertEquals bug, add test suite
The CMake build now includes all required sections for
 TestCPP_Exceptions_test, which is the addition of the Exceptions test
 suite to the build.
Now a TestCPP_Exceptions_test executable is created and linked in.

Added a test suite for TestCPPExceptions.
This included adding the following files:
 - test/include/Exceptions/ExceptionsTests.h
   - Test declarations
 - test/include/Exceptions/ExceptionsSuite.h
   - Creation of test suite
 - test/src/Exceptions/ExceptionsTests.cpp
   - Implementation of tests
 - test/src/TestCPPExceptionsMain.cpp
   - Implementation of suite runner
The only things that really needed to happen here were:
 - Test constructor variants
 - Make sure the constructors don't throw when given sane output

Additional behavior to check:
 - What does it do with negative testing?

If everything is good with CI after I push this I'll add some negative
 testing.

In Assertions, assertEquals did not work correctly with C Strings.
I added a template specialization that fixes this... I have a feeling
 this is going to need to happen in other cases as well, though I don't
 yet know what.