Qualify forward with std:: like move, fix shadowing, fix #5, formatting
Whitespace/newline formatting in function header.

Adjust the name of the TestSuite ctor parameter so it doesn't shadow
 the member field.

Create a common initialization function since the majority of the
 initialization logic is the same in both cases and I added a few more
 initialization statements; it is more correct to have them initialized
 by the constructor than the run() function for the initial suite run.
I added a firstRun bool to indicate when the run() re-initialization
 logic should be run since the suites can be rerun if desired; for
 example if the client application is running tests interactively and
 the end-user wants to run a suite 4 or 5 times to validate that tests
 are not flakey, or if an automated application wants to do something
 along those lines, or with other use cases that I have not conceived.
The common initialization function is inlined and removed by the
 compiler, which is a warning in MSVC; it is suppressed with a #pragma
 since it is an internal function, and it's ok if it is removed because
 of inlining with no further references.
Maybe using std::forward helped the compiler to decide to inline? Or
 maybe it would have done it anyways.

Removed using std::forward since it is now fully-qualified.