Convert Boost.Stacktrace and transitives to git submodules and related
Added new SubmodulesInit include to top-level CMakeLists.txt.
This ensures CMake updates all the dependencies that are git submodules
 so everything is at the right tag/commit and exists.
SubmodulesInit.cmake is fully implemented and derived from
https://cliutils.gitlab.io/modern-cmake/chapters/projects/submodule.html

No need to conditionally worry about including from Boost.Stacktrace
 since it is now handled through the git submodule dependencies.
Includes.cmake is altered to take this into account.

Add the required boost_* targets to the install target to ensure the
 necessary files are there to support the library when TestCPP gets
 installed.
These targets are identified and added to the list in
 Boost.Stacktrace.cmake.
Formatting.

Update the TestCPP package email contact to the appropriate HSF email
 address.

Implement a more selective and proper Boost.Stacktrace.cmake.
This version is easier to maintain if the Boost.Stacktrace dependencies
 change in the future.
It also lays out what targets to install given the environment we're
 building in.